LineConverter is done.
In my development, I have come to realize that there is a large problem
with the way RenderEngine only takes static objects and not temporary
ones. It is impossible to render a single object for just one frame.
Also it is difficult to manage the camera for GUI rendering. It would
all be so simple is I was not so adamant on instancing.