Make RenderEngine use RenderContext's Camera
I removed the Render class. The camera system for perspective is a bit
of a mess right now. The problem is that I need a camera, but have
nowhere to get the settings for it. I also am working on a new Model.
I'm not sure if I should keep it though. It needs a lot of work for a
supplementary modeling framework. Eventually I will need a modeling
framework so I should probably keep it. It will definitely be on the
back burner for a while.