Fix Texture, Add ModelGroup.
While working on converting the gui rendering over to just a regular
model, I found out some of the shortcomings of Texture. It also felt
natural to make groups for models separate of a RenderContext. It also
felt natural to make RenderContext have a group of it's own because of
all the repeated code. This eliminates repeated code and makes it more
powerful in general. Another small thing I did while fixing Texture is
change how the OpenGL render engine gets texels from nearest to linear.