Added a follow camera
Right now it is a pretty terrible velocity-pointing follow camera. I
think what I must do in the future is make the camera keep a finite
history of past locations and average those out. This means that there
will be no jerky movement and the center will always point at the
character. It also means that because the camera is still angled it will
still look forward.